fix: 审查修复——通义千问API请求格式 + ASR错误消息退化
根据代码审查报告(IDE-171)修复2个关键问题：
1. 通义千问 API 请求体格式改为 DashScope 原生格式
   - messages 包装进 input，temperature/max_tokens 包装进 parameters
2. ASR 错误消息退化修复
   - C++ 端优先读取 message（中文提示），fallback 到 error（短代码）
3. GStreamer signal 参数类型规范化 (int)1 → (gdouble)1.0

+++ 06 AI工作提效/40 UOS 速记260727/src/audio/audiorecorder.cpp
@@ -224,7 +224,7 @@ void AudioRecorder::timerEvent(QTimerEvent *event)
     // 从 GStreamer level 元素获取实时音频电平
     if (m_level) {
         GstStructure *s = nullptr;
-        g_signal_emit_by_name(m_level, "get-level", (int)1, 0.0, &s);
+        g_signal_emit_by_name(m_level, "get-level", (gdouble)1.0, 0.0, &s);
         if (s) {
             gdouble peak_dB;
             if (gst_structure_get_double(s, "peak", &peak_dB)) {
